moving nav related code inside the header
