+++ frontend/src/components/Header.js
@@ -1,11 +1,54 @@
 import React, { Component } from 'react';
 // add library
 import styled from 'styled-components'
-import { SearchBar, Section, ListItem, Logo, SideList, Footer } from '../components';
+import { SearchBar, Section, ListItem, Logo, SideList } from '../components';
 import { user } from '../resources/images';
 import { blue } from '../resources/colors';
+import { graphql, compose } from 'react-apollo';
+import gql from 'graphql-tag';
+import Modal from 'react-modal';
+import Login from '../screens/Login';
 
-export default class Header extends Component {
+class Header extends Component {
+    state = {
+      name: '',
+      email: '',
+      password: '',
+      error: '',
+      openModal: false,
+    };
+
+    onSignup() {
+  		const { name, email, password } = this.state;
+  		this.props
+  			.mutate({
+  				variables: { name, email, password },
+  			})
+  			.then(data => console.log(JSON.stringify(data)))
+  			.catch(err => console.log(err, password, email, name));
+  	}
+
+  	onLogin() {
+  		const { email, password } = this.state;
+  		if (email.length === 0 || password.length === 0) {
+  			this.setState({ error: 'Empty fields' });
+  		} else {
+  			this.setState({ error: '' });
+  			this.props
+  				.mutate({
+  					variables: { email, password },
+  				})
+  				.then(data => {
+  					console.log(JSON.stringify(data));
+  					this.setState({ openModal: false });
+  				})
+  				.catch(error => {
+  					console.log(error);
+  					this.setState({ error: 'Wrong email/password ' });
+  				});
+  		}
+  	}
+
     render() {
         return (
           <HeaderContainer>
@@ -15,17 +58,62 @@ export default class Header extends Component {
             <Section title="Login" onClick={() => this.setState({ openModal: true })} />
             <Section title="Signup" onClick={() => this.onSignup()} />
             <img src={user} style={avatar} alt={user} />
-          </HeaderContainer>
 
+            <Modal
+    					style={{
+    						overlay: {
+    							backgroundColor: 'rgba(255,255,255,.2)',
+    							alignItems: 'center',
+    							justifyContent: 'center',
+    						},
+    						content: {
+    							backgroundColor: 'transparent',
+    							borderWidth: 0,
+    							padding: 50,
+    						},
+    					}}
+    					isOpen={this.state.openModal}
+    					onRequestClose={() => this.setState({ openModal: false })}
+    					contentLabel="Modal with image"
+    				>
+    					<Login
+    						handleEmail={event => this.setState({ email: event.target.value })}
+    						handlePassword={event => this.setState({ password: event.target.value })}
+    						onClick={() => this.onLogin()}
+    						error={this.state.error}
+    					/>
+    				</Modal>
+          </HeaderContainer>
         )
     }
 }
+
+const SIGNUP = gql`
+	mutation Signup($name: String!, $email: String!, $password: String!) {
+		signup(name: $name, email: $email, password: $password) {
+			_id
+		}
+	}
+`;
+
+const LOGIN = gql`
+	mutation Login($email: String!, $password: String!) {
+		login(email: $email, password: $password) {
+			_id
+			name
+			email
+			rememberToken
+		}
+	}
+`;
+
+export default graphql(LOGIN, SIGNUP)(Header);
 
 // styled component
 const HeaderContainer = styled.div`
 	display: flex;
 	flex-direction: row;
-	background: #070a37;
+	background: ${blue};
 	height: 85px;
 	align-items: center;
 `
